[refactor] : 상품 관련 엔티티 그룹핑
