+++ src/main/java/shop/itbook/itbookshop/orderproduct/entity/OrderProduct.java
@@ -15,9 +15,8 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import org.hibernate.annotations.ColumnDefault;
 import shop.itbook.itbookshop.order.entity.Order;
-import shop.itbook.itbookshop.product.entity.Product;
+import shop.itbook.itbookshop.productgroup.product.entity.Product;
 
 /**
  * 주문상품에 대한 엔티티입니다.
