+++ src/main/java/shop/itbook/itbookshop/productgroup/productrelationgroup/entity/ProductRelationGroup.java
@@ -1,0 +1,36 @@
+package shop.itbook.itbookshop.productgroup.productrelationgroup.entity;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+/**
+ * 연관 상품 테이블에 대한 엔티티 입니다.
+ *
+ * @author 강명관
+ * @since 1.0
+ */
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+@Entity
+@Table(name = "product_relation_group")
+public class ProductRelationGroup {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "product_relation_group_no")
+    private Integer productRelationGroupNo;
+
+    @Column(name = "product_relation_group_name", nullable = false,
+        columnDefinition = "varchar(255)", unique = true)
+    private String productRelationGroupName;
+}
