+++ src/main/java/shop/itbook/itbookshop/productgroup/productrelationgroupreservation/entity/ProductRelationGroupReservation.java
@@ -1,7 +1,6 @@
-package shop.itbook.itbookshop.cart.entity;
+package shop.itbook.itbookshop.productgroup.productrelationgroupreservation.entity;
 
 import java.io.Serializable;
-import javax.persistence.Column;
 import javax.persistence.Embeddable;
 import javax.persistence.EmbeddedId;
 import javax.persistence.Entity;
@@ -15,11 +14,11 @@ import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import shop.itbook.itbookshop.member.entity.Member;
 import shop.itbook.itbookshop.productgroup.product.entity.Product;
+import shop.itbook.itbookshop.productgroup.productrelationgroup.entity.ProductRelationGroup;
 
 /**
- * 장바구니 테이블에 대한 엔티티 입니다.
+ * 상품과 연관 상품 그룹 관계 테이블에 대한 엔티티 입니다.
  *
  * @author 강명관
  * @since 1.0
@@ -29,27 +28,24 @@ import shop.itbook.itbookshop.productgroup.product.entity.Product;
 @AllArgsConstructor
 @NoArgsConstructor
 @Entity
-@Table(name = "cart")
-public class Cart {
+@Table(name = "product_relation_group_reservation")
+public class ProductRelationGroupReservation {
 
     @EmbeddedId
     private Pk pk;
 
-    @MapsId("memberNo")
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_no")
-    private Member member;
-
     @MapsId("productNo")
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "product_no")
+    @JoinColumn(name = "product_no", nullable = false)
     private Product product;
 
-    @Column(name = "product_count", columnDefinition = "integer default 1")
-    private Integer productCount;
+    @MapsId("productRelationGroupNo")
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "product_relation_group_reservation_no", nullable = false)
+    private ProductRelationGroup productRelationGroup;
 
     /**
-     * The type Pk. 회원과 상품에 대한 복합키 설정을 위한 클래스 입니다.
+     * The type Pk. 관계 테이블의 복합키를 주키로 하기 위한 클래스 입니다.
      *
      * @author 강명관
      * @since 1.0
@@ -61,10 +57,10 @@ public class Cart {
     @Embeddable
     public static class Pk implements Serializable {
 
-        private Long memberNo;
+        private Integer productNo;
 
-        private Long productNo;
+        private Long productRelationGroupNo;
 
     }
-}
 
+}
