+++ src/main/java/shop/itbook/itbookshop/productgroup/tag/entity/Tag.java
@@ -1,0 +1,36 @@
+package shop.itbook.itbookshop.productgroup.tag.entity;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+/**
+ * 상품 태그 테이블에 대한 엔티티 입니다.
+ *
+ * @author 강명관
+ * @since 1.0
+ */
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+@Entity
+@Table(name = "tag")
+public class Tag {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "tag_no")
+    private Integer tagNo;
+
+    @Column(name = "tag_name", nullable = false, unique = true, columnDefinition = "varchar(20)")
+    private String tagName;
+}
