--- src/main/java/shop/itbook/itbookshop/productimage/entity/ProductImage.java
@@ -1,43 +1,0 @@
-package shop.itbook.itbookshop.productimage.entity;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import shop.itbook.itbookshop.product.entity.Product;
-
-/**
- * 상품 이미지 테이블에 대한 엔티티 입니다.
- *
- * @author 강명관
- * @since 1.0
- */
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-@Entity
-@Table(name = "product_image")
-public class ProductImage {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "product_image_no")
-    private Long productImageNo;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "product_no", nullable = false)
-    private Product product;
-
-    @Column(name = "image_url", nullable = false, columnDefinition = "text")
-    private String imageUrl;
-}
